@@ -54,12 +54,11 @@
     ]
   },
   "peerDependencies": {
-    "@material-ui/core": "^4.0.0",
-    "@material-ui/icons": "^4.0.0",
-    "@material-ui/pickers": "^3.3.0",
-    "@material-ui/styles": "^4.0.0",
-    "react": "^16.8.0 || ^17.0.0",
-    "react-dom": "^16.8.0 || ^17.0.0"
+    "@mui/icons-material": "^5.0.0",
+    "@mui/lab": "^5.0.0-alpha.30",
+    "@mui/material": "^5.0.0",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0"
   },
   "dependencies": {
     "dot-prop": "^6.0.1",
