@@ -1,5 +1,4 @@
-import { alpha, createStyles, makeStyles, TableRow } from "@material-ui/core";
-import clsx from "clsx";
+import { alpha, css, SxProps, TableRow, Theme } from "@mui/material";
 import PropTypes from "prop-types";
 import React, { MouseEventHandler, PropsWithChildren, useCallback, useContext, useMemo } from "react";
 import TableContext, { TableState } from "../table.context";
@@ -14,26 +13,10 @@ interface BodyRowProps<RowType extends BaseData> {
   data: RowType;
 }
 
-const useStyles = makeStyles(
-  (theme) =>
-    createStyles({
-      alternateRowColour: {
-        backgroundColor: alpha(theme.palette.error.dark, 0.9),
-      },
-      disabledRow: {
-        backgroundColor: theme.palette.action.disabledBackground,
-        opacity: theme.palette.action.disabledOpacity,
-      },
-      rowHover: {
-        cursor: "pointer",
-        backgroundColor: theme.palette.action.hover,
-      },
-      selectedRow: {
-        backgroundColor: theme.palette.action.selected,
-      },
-    }),
-  { name: "DataTable-BodyRow" },
-);
+const rowHoverClass = css`
+  cursor: "pointer";
+  background-color: "palette.action.hover";
+`;
 
 /**
  * The BodyRow component is a single row in the table body and is responsible for managing the state
@@ -45,38 +28,31 @@ const useStyles = makeStyles(
 const BodyRow = <RowType extends BaseData, AllDataType extends RowType[]>({
   index,
   data,
-  ...props
 }: PropsWithChildren<BodyRowProps<RowType>>) => {
-  const classes = useStyles(props);
   const { structure, rowClick, rowOptions, rowsSelectable, selectedRows } =
     useContext<TableState<RowType, AllDataType>>(TableContext);
 
   const rowId = useMemo(() => getRowId(data, index), [data, index]);
   const isAlternateRowColour = useMemo(() => rowOptions?.alternateRowColour?.(data) ?? false, [data, rowOptions]);
   const isDisabledRow = useMemo(() => rowOptions?.rowDisabled?.(data) ?? false, [data, rowOptions]);
-  const tableRowClasses = useMemo(
-    () =>
-      clsx({
-        [classes.alternateRowColour]: isAlternateRowColour,
-        [classes.disabledRow]: isDisabledRow,
-        [classes.selectedRow]: rowsSelectable && Boolean(selectedRows[rowId]),
-      }),
-    [
-      classes.alternateRowColour,
-      classes.disabledRow,
-      classes.selectedRow,
-      isAlternateRowColour,
-      isDisabledRow,
-      rowId,
-      rowsSelectable,
-      selectedRows,
+  const tableRowStyles = useMemo<SxProps<Theme>>(
+    () => [
+      !isAlternateRowColour ? {} : (theme) => ({ backgroundColor: alpha(theme.palette.error.dark, 0.9) }),
+      !isDisabledRow
+        ? {}
+        : (theme) => ({
+            backgroundColor: theme.palette.action.disabledBackground,
+            opacity: theme.palette.action.disabledOpacity,
+          }),
+      !rowsSelectable || !selectedRows[rowId] ? {} : (theme) => ({ backgroundColor: theme.palette.action.selected }),
     ],
+    [isAlternateRowColour, isDisabledRow, rowId, rowsSelectable, selectedRows],
   );
 
   const onHover = useCallback<MouseEventHandler<HTMLTableRowElement>>(
     (e) => {
       if (!rowClick) return;
-      e.currentTarget.classList.add(classes.rowHover);
+      e.currentTarget.classList.add(rowHoverClass.name);
       if (!e.currentTarget.parentNode) return;
       const hoverRowIndex = Array.from(e.currentTarget.parentNode.children).indexOf(e.currentTarget);
       for (let row = e.currentTarget.previousSibling as Element; row; row = row.previousSibling as Element) {
@@ -85,22 +61,22 @@ const BodyRow = <RowType extends BaseData, AllDataType extends RowType[]>({
         const rowsBetween = hoverRowIndex - rowIndex;
         row.querySelectorAll("td[rowspan]").forEach((cell) => {
           const rowSpan = Number(cell.getAttribute("rowspan"));
-          if (rowSpan > rowsBetween) cell.classList.add(classes.rowHover);
+          if (rowSpan > rowsBetween) cell.classList.add(rowHoverClass.name);
         });
       }
     },
-    [classes.rowHover, rowClick],
+    [rowClick],
   );
 
   const onUnHover = useCallback<React.MouseEventHandler<HTMLTableRowElement>>(
     (e) => {
       if (!rowClick) return;
-      e.currentTarget.classList.remove(classes.rowHover);
+      e.currentTarget.classList.remove(rowHoverClass.name);
       for (let row = e.currentTarget.previousSibling as Element; row; row = row.previousSibling as Element) {
-        row.querySelectorAll("td[rowspan]").forEach((cell) => cell.classList.remove(classes.rowHover));
+        row.querySelectorAll("td[rowspan]").forEach((cell) => cell.classList.remove(rowHoverClass.name));
       }
     },
-    [classes.rowHover, rowClick],
+    [rowClick],
   );
 
   const bodyContextValue = useMemo(
@@ -114,13 +90,7 @@ const BodyRow = <RowType extends BaseData, AllDataType extends RowType[]>({
   );
 
   return (
-    <TableRow
-      key={rowId}
-      data-testid="tableRow"
-      onMouseOver={onHover}
-      onMouseOut={onUnHover}
-      className={tableRowClasses}
-    >
+    <TableRow key={rowId} data-testid="tableRow" onMouseOver={onHover} onMouseOut={onUnHover} sx={tableRowStyles}>
       {structure.flattened.notHidden.map((struct) => (
         <BodyContextProvider
           key={struct.key}
